+++ app/src/main/java/top/yogiczy/mytv/BootReceiver.kt
@@ -14,7 +14,7 @@ class BootReceiver : BroadcastReceiver() {
     override fun onReceive(context: Context, intent: Intent) {
         if (Intent.ACTION_BOOT_COMPLETED == intent.action) {
             val sp: SharedPreferences = SP.getInstance(context)
-            val bootLaunch = sp.getBoolean(SP.KEY.APP_BOOT_LAUNCH.name, false)
+            val bootLaunch = sp.getBoolean("APP_BOOT_LAUNCH", false)
 
             if (bootLaunch) {
                 context.startActivity(Intent(context, LeanbackActivity::class.java).apply {