+++ app/src/main/java/top/yogiczy/mytv/ui/screens/leanback/video/player/Media3VideoPlayer.kt
@@ -109,8 +109,6 @@ class LeanbackMedia3VideoPlayer(
                         prepare(uri, C.CONTENT_TYPE_HLS)
                     } else if (contentTypeAttempts[C.CONTENT_TYPE_OTHER] != true) {
                         prepare(uri, C.CONTENT_TYPE_OTHER)
-                    } else if (contentTypeAttempts[C.CONTENT_TYPE_OTHER] != true) {
-                        prepare(uri, C.CONTENT_TYPE_OTHER)
                     } else {
                         triggerError(PlaybackException.UNSUPPORTED_TYPE)
                     }